move the convenience initializers in a separate file
